about %ld

+++ src/udp_ssr_client.c
@@ -188,7 +188,7 @@ void client_udp_remote_recv_cb(uv_udp_t* handle, ssize_t nread, const uv_buf_t* 
         pr_info("[udp] session %s <=> %s recv remote data length %ld",
             get_addr_str(&remote_ctx->incoming_addr.addr, tmp1, sizeof(tmp1)),
             get_addr_str(&remote_ctx->target_addr.addr, tmp2, sizeof(tmp2)),
-            buffer_get_length(buf));
+            (long)buffer_get_length(buf));
 
         err = ss_decrypt_all(listener_ctx->cipher_env, buf, (size_t)nread*2);
         if (err) {
@@ -468,7 +468,7 @@ client_udp_listener_recv_cb(uv_udp_t* handle, ssize_t nread, const uv_buf_t* uvb
 
             pr_info(remote_ctx ? "[udp] session %s <=> %s reused, data length %ld" : "[udp] session %s <=> %s starting, data length %ld",
                 get_addr_str(addr, tmp1, sizeof(tmp1)),
-                get_addr_str(&target_addr.addr, tmp2, sizeof(tmp2)), (size_t)nread);
+                get_addr_str(&target_addr.addr, tmp2, sizeof(tmp2)), (long)nread);
 
             if (remote_ctx == NULL) {
                 remote_ctx = create_client_udp_remote(loop, listener_ctx->timeout, client_udp_remote_timeout_cb);
